+++ src/game.h
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <SDL.h>
-#include "utils/int2.hpp"
-#include "utils/float2.hpp"
+#include "int2.hpp"
+#include "float2.hpp"
 #include "engine/camera.h"
 
 namespace Tmpl8 {
