+++ tmpl_2019-08.vcxproj.filters
@@ -15,7 +15,7 @@
     <ClCompile Include="src\graphics\tilemap.cpp" />
     <ClCompile Include="src\game\enemy.cpp" />
     <ClCompile Include="src\game\entity.cpp" />
-    <ClCompile Include="src\utils\struct\arena.cpp" />
+    <ClCompile Include="src\game\arena\arena.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="src\game.h" />
@@ -38,7 +38,7 @@
     <ClInclude Include="src\game\entity.h" />
     <ClInclude Include="src\utils\types.h" />
     <ClInclude Include="src\graphics\entitymap.h" />
-    <ClInclude Include="src\utils\struct\arena.h" />
+    <ClInclude Include="src\game\arena\arena.h" />
   </ItemGroup>
   <ItemGroup>
     <Filter Include="template code">
